requirements.txt
fastapi
uvicorn
psycopg2-binary
python-multipart
numpy
scipy
librosa
noisereduce
soundfile